+++ cases/cloud.mail - folder_test/create_folder_inside_folder_test.js
@@ -1,5 +1,5 @@
-import account from "../../../steps/account";
-import folders from "../../../steps/folders";
+import account from "../../steps/account";
+import folders from "../../steps/folders";
 
 describe('Работа с папками', () => {
